+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * feat: add `hasAskForPermissionsConsentCard` for requesting permissions
 * feat: add `hasLinkAccountCard` for account linking
 * fix: correct standard card error messages to say "standard card"
+* feat: support all UpsServiceClient calls by adding `distanceUnits`, `temperatureUnits`, and `timeZone` to `withProfile`
 
 ## 2.4.0
 * update libraries